add test component to avoid multiple fetch being invoked
